Extract generate form code command to a new package
