+++ packages/bpmn-vscode-extension/package.json
@@ -30,10 +30,7 @@
     "@kie-tools-core/i18n": "workspace:*",
     "@kie-tools-core/patternfly-base": "workspace:*",
     "@kie-tools-core/vscode-extension": "workspace:*",
-    "@kie-tools/form-code-generator": "workspace:*",
-    "@kie-tools/form-code-generator-bootstrap4-theme": "workspace:*",
-    "@kie-tools/form-code-generator-patternfly-theme": "workspace:*",
-    "@kie-tools/jbpm-form-code-generator-themes": "workspace:*",
+    "@kie-tools/form-code-generator-vscode-command": "workspace:*",
     "@kie-tools/kie-bc-editors": "workspace:*"
   },
   "devDependencies": {
