+++ packages/form-code-generator-vscode-command/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@kie-tools/tsconfig/tsconfig.esm.json"
+}
