+++ repo/graph.json
@@ -73,10 +73,9 @@
       { "id": "@kie-tools/xml-parser-ts-codegen" },
       { "id": "@kie-tools/xml-parser-ts" },
       { "id": "bpmn-vscode-extension" },
-      { "id": "@kie-tools/jbpm-form-code-generator-themes" },
+      { "id": "@kie-tools/form-code-generator-vscode-command" },
       { "id": "@kie-tools/vscode-java-code-completion-extension-plugin" },
-      { "id": "@kie-tools/form-code-generator-bootstrap4-theme" },
-      { "id": "@kie-tools/form-code-generator-patternfly-theme" },
+      { "id": "@kie-tools/jbpm-form-code-generator-themes" },
       { "id": "@kie-tools/pmml-editor-marshaller" },
       { "id": "@kie-tools/stunner-editors-dmn-loader" },
       { "id": "chrome-extension-pack-kogito-kie-editors" },
@@ -136,6 +135,8 @@
       { "id": "@kie-tools/extended-services-java" },
       { "id": "extended-services-vscode-extension" },
       { "id": "@kie-tools/form-code-generator" },
+      { "id": "@kie-tools/form-code-generator-bootstrap4-theme" },
+      { "id": "@kie-tools/form-code-generator-patternfly-theme" },
       { "id": "@kie-tools/import-java-classes-component" },
       { "id": "@kie-tools/runtime-tools-process-webapp-components" },
       { "id": "@kie-tools/runtime-tools-shared-webapp-components" },
@@ -898,7 +899,7 @@
       },
       {
         "source": "bpmn-vscode-extension",
-        "target": "@kie-tools/jbpm-form-code-generator-themes",
+        "target": "@kie-tools/form-code-generator-vscode-command",
         "weight": 1
       },
       {
@@ -917,13 +918,8 @@
         "weight": 1
       },
       {
-        "source": "@kie-tools/jbpm-form-code-generator-themes",
-        "target": "@kie-tools/form-code-generator-bootstrap4-theme",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools/jbpm-form-code-generator-themes",
-        "target": "@kie-tools/form-code-generator-patternfly-theme",
+        "source": "@kie-tools/form-code-generator-vscode-command",
+        "target": "@kie-tools/jbpm-form-code-generator-themes",
         "weight": 1
       },
       {
@@ -932,23 +928,13 @@
         "weight": 1
       },
       {
-        "source": "@kie-tools/form-code-generator-bootstrap4-theme",
-        "target": "@kie-tools/form-code-generator",
+        "source": "@kie-tools/jbpm-form-code-generator-themes",
+        "target": "@kie-tools/form-code-generator-bootstrap4-theme",
         "weight": 1
       },
       {
-        "source": "@kie-tools/form-code-generator-bootstrap4-theme",
-        "target": "@kie-tools-core/webpack-base",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools/form-code-generator-patternfly-theme",
-        "target": "@kie-tools/form-code-generator",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools/form-code-generator-patternfly-theme",
-        "target": "@kie-tools-core/webpack-base",
+        "source": "@kie-tools/jbpm-form-code-generator-themes",
+        "target": "@kie-tools/form-code-generator-patternfly-theme",
         "weight": 1
       },
       {
@@ -1674,6 +1660,26 @@
       {
         "source": "@kie-tools/form-code-generator",
         "target": "@kie-tools/tsconfig",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/form-code-generator-bootstrap4-theme",
+        "target": "@kie-tools/form-code-generator",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/form-code-generator-bootstrap4-theme",
+        "target": "@kie-tools-core/webpack-base",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/form-code-generator-patternfly-theme",
+        "target": "@kie-tools/form-code-generator",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/form-code-generator-patternfly-theme",
+        "target": "@kie-tools-core/webpack-base",
         "weight": 1
       },
       {
@@ -2647,6 +2653,7 @@
     ["@kie-tools/form-code-generator", "packages/form-code-generator"],
     ["@kie-tools/form-code-generator-bootstrap4-theme", "packages/form-code-generator-bootstrap4-theme"],
     ["@kie-tools/form-code-generator-patternfly-theme", "packages/form-code-generator-patternfly-theme"],
+    ["@kie-tools/form-code-generator-vscode-command", "packages/form-code-generator-vscode-command"],
     ["@kie-tools/form-dmn", "packages/form-dmn"],
     ["@kie-tools-core/i18n", "packages/i18n"],
     ["@kie-tools/i18n-common-dictionary", "packages/i18n-common-dictionary"],
